--- a/lib/graphql/schema/relay_classic_mutation.rb
+++ b/lib/graphql/schema/relay_classic_mutation.rb
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require "graphql/types/string"
 
 module GraphQL
   class Schema