--- a/lib/graphql/types.rb
+++ b/lib/graphql/types.rb
@@ -1,12 +1,19 @@
 # frozen_string_literal: true
-require "graphql/types/boolean"
-require "graphql/types/big_int"
-require "graphql/types/float"
-require "graphql/types/id"
-require "graphql/types/int"
-require "graphql/types/iso_8601_date"
-require "graphql/types/iso_8601_date_time"
-require "graphql/types/iso_8601_duration"
-require "graphql/types/json"
-require "graphql/types/string"
-require "graphql/types/relay"
+
+module GraphQL
+  module Types
+    extend Autoload
+
+    autoload :Boolean, "graphql/types/boolean"
+    autoload :BigInt, "graphql/types/big_int"
+    autoload :Float, "graphql/types/float"
+    autoload :ID, "graphql/types/id"
+    autoload :Int, "graphql/types/int"
+    autoload :JSON, "graphql/types/json"
+    autoload :String, "graphql/types/string"
+    autoload :ISO8601Date, "graphql/types/iso_8601_date"
+    autoload :ISO8601DateTime, "graphql/types/iso_8601_date_time"
+    autoload :ISO8601Duration, "graphql/types/iso_8601_duration"
+    autoload :Relay, "graphql/types/relay"
+  end
+end