--- a/spec/fixtures/eager_module/eager_class.rb
+++ b/spec/fixtures/eager_module/eager_class.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module EagerModule
+  module EagerClass
+  end
+end