--- a/spec/fixtures/lazy_module/lazy_class.rb
+++ b/spec/fixtures/lazy_module/lazy_class.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module LazyModule
+  module LazyClass
+  end
+end