Reduces require time from ~160ms to ~5ms. Autoloads all constants under `GraphQL` to optimize boot time. Also autoloads `GraphQL::Types` and `GraphQL::Query` because they can't be required in sequence.

Script:
```rb
#!/usr/bin/env ruby
# frozen_string_literal: true

require "bundler/setup"
require "benchmark"

ms = 1000 * Benchmark.realtime do
  require "graphql"
end

puts "#{ms}ms"
```

Before:

```
> bin/boot
169.55200000666082ms
```

After:

```
> bin/boot
5.60200004838407ms
```

Rails will automatically eager load GraphQL, but other frameworks / setups will need to call `GraphQL.eager_load!`. If we agree with the approach, I can document this. Because we don't require on Active Support, we need to manually implement eager loading. `ActiveSupport::Autoload` implements the `eager_load!` method for you.

Also, if you're interested, I can implement testing for eager loading (example), but we'll need to roll our own test isolation feature.