@rmosolgo 2024-11-29
Hey, thanks so much for putting in the time on this. It's been in the back of my mind for a long time to improve the organization of all those `require`s, but I didn't think of approaching it this way. 


> other frameworks / setups will need to call `GraphQL.eager_load!` 

This seems alright to me. I'm trying to think if we can detect these and warn in cases where this should be called. Are there other setups that come to mind besides Sinatra and Hanami? (I mean, there's plain Rack, and a long tail of other possibilities, but any other big ones?)


>  testing for eager loading


I'm comfortable assuming that `autoload` works... but is there any way that we could make sure `eager_load!` will stay in-sync with our `autoload ...` configurations? For example, something like: 

```ruby 
all_constants = GraphQL.constants.map { |const_name| GraphQL.const_get(const_name) }
eager_loadable_constants = all_constants - [
  # Any specific exceptions to autoloading can go here, along with a comment for why they're skipped
]
assert_equal eager_loadable_constants, GraphQL.eager_load!
```

And then do something similar for `GraphQL::Query`. 

Yes, those other tests are failing on `master` 😿 . I haven't been able to get the `system_tests` to fail locally and I can't figure out how to fix the flaky test where database tables are somehow missing...