@gmcgibbon 2024-11-29
> Are there other setups that come to mind besides Sinatra and Hanami?

I'll investigate this. Let's narrow the scope to Sinatra, Hanami, and Rails. Everything else can be supported later, and we can instruct folks to call that method themselves in more exotic cases.

> I'm comfortable assuming that autoload works... but is there any way that we could make sure eager_load! will stay in-sync with our autoload ... configurations?

Sounds good. I'll re-implement a minimal version of `ActiveSupport::Autload` so that we can automate the eager_load! method.