@gmcgibbon 2024-11-29
I wrote a document page, added `GraphQL::Autoload`, added a warning for production not eager loading, and some rudimentary autoload tests. I can test the warning, but that would require an isolated test suite and re-implementing `ActiveSupport::Testing::Isolation`.