@rmosolgo 2024-12-02
I'm looking over it one more time and adding more docs... it seems like `eager_load!` should continue recursively, calling `.eager_load!` on any loaded constants which respond to that method (eg, `GraphQL.eager_load!` should call `GraphQL::Query.eager_load!`). Does that sound right to you? I'll open a PR with the change to review.