@sealabcore 2024-12-05
We are seeing this new error in our Rails app on web server boot, production console boot, and some rake tasks. Should we be seeing this? The docs in this PR mention that nothing should be needed when using the Graphql-Ruby gem with a Rails app.

<img width="860" alt="Screenshot 2024-12-04 at 1 49 28 PM" src="https://github.com/user-attachments/assets/9f029ba3-0276-4643-8c4c-ec011406b6e0">

<img width="962" alt="Screenshot 2024-12-04 at 4 28 11 PM" src="https://github.com/user-attachments/assets/acec9f71-3f4e-4a3f-acee-29b28f87624f">
<img width="987" alt="Screenshot 2024-12-04 at 4 28 26 PM" src="https://github.com/user-attachments/assets/51335663-194d-435d-a9cb-42f1d7e008f4">
<img width="1077" alt="Screenshot 2024-12-04 at 4 34 05 PM" src="https://github.com/user-attachments/assets/49e51c6d-7ca4-43f9-acf7-dd1001c9a28e">