@rmosolgo 2024-12-05
Hey @sealabcore, no... it's not supposed to show that all the time :S I was able to replicate this immediately with `rails new ...`.  Digging in, its a problem with the _warning_ -- its detection for whether autoloading is happening or not is not sophisticated enough. 

First, Rails calls `Zeitwerk::Loader.eager_load_all`, which loads application code (including `app/graphql/my_schema.rb`), which references `GraphQL::Schema`, which emits this warning. 

Then, it does `eager_load_namespaces.each(&:eager_load!)` which eager-loads GraphQL's constants. 

Here's the source for that: https://github.com/rails/rails/blob/91d456366638ac6c3f6dec38670c8ada5e7c69b1/railties/lib/rails/application/finisher.rb#L79-L81

So I think everything _works_ -- everything is loaded when the application starts -- but the integration needs to be improved. I'll take a closer look tomorrow, sorry for the trouble!