@gmcgibbon 2024-12-05
Hm, the application could reference any GraphQL constant in theory, so it might make more sense to use `before_eager_load { GraphQL.eager_load! }` instead of appending to `eager_load_namesapces`.