@gmcgibbon 2024-12-05
I can confirm this fixes the issue. If we want to make the the gem is eager loaded before we eager load the app, I think this is our best bet, so I'll issue the patch. Sorry for the issues folks.