@rmosolgo 2024-12-05
Thanks for taking care of it, @gmcgibbon! I just shipped that patch in 2.4.6. 

@sealabcore, thanks for reporting this and sorry for the trouble. Please give that new version a try and let me know if it gives you any more issues!