@sealabcore 2024-12-05
Thanks to you both for the quick resolution! I’ll give this a try today.