@zvkemp 2024-12-09
We've had a report that this change breaks constant loading if any of the nested files are required during app boot (I can open a new issue on this if necessary).

Here's a minimal reproduction:

```ruby
require 'bundler/inline'

gemfile do
  source 'https://rubygems.org'
  gem 'graphql', '2.4.7'
end

require 'graphql'
require 'graphql/tracing/active_support_notifications_trace'

GraphQL.eager_load!
```

What I think is happening is that the nested graphql/tracing/**.rb files don't require their parent (tracing.rb). So requiring this file:
 - calls `require "graphql/tracing/platform_trace"`, 
 - which loads platform_trace.rb, 
 - which encounters the autoload-registered constant `GraphQL::Tracing`
 - which calls `require 'graphql/tracing'
 - which calls `require 'graphql/tracing/platform_trace'`, which is skipped because this file has already been required — at this point the loading of the `PlatformTrace` module is higher up in the callstack, so it's not going to complete until after the rest of tracing.rb is loaded
 - Unfortunately, the rest of tracing.rb includes files that refer to PlatformTrace, so this error is raised:
`NameError: uninitialized constant GraphQL::Tracing::AppOpticsTrace::PlatformTrace`

I believe all of these nested files should include a top-level `require 'graphql/tracing'` to ensure the order is correct.